Add DOCTYPE HTML boilerplate and h1 heading
Created a basic HTML boilerplate with a DOCTYPE declaration. Added an h1 heading to the HTML body with the text World's Best Lasagna! This sets up the initial structure of the webpage and provides a title for the content to follow. Plus it gets you hyped for the best lasagna ever seen!